pyproject.toml
[project]
name = "social-engine"
version = "0.1.0"
description = "Minimal content capture and Publer scheduling tool."
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0.0",
  "requests>=2.31.0",
]

[project.scripts]
social-engine = "src.cli:main"
